+++ src/actions/actions.js
@@ -1,5 +1,3 @@
-// import fetch from 'cross-fetch';
-
 export const SIGN_IN = 'SIGN_IN';
 export const SIGN_UP = 'SIGN_UP';
